Validate middlewares in Validate()
